set up runnum epX check

+++ analysis_scripts/asymmetry_extraction/fitting_process.cpp
@@ -784,7 +784,8 @@ TH1D* createHistogramForBin_single_hadron(const char* histName, int binIndex,
   TTreeReaderValue<double> beam_pol(dataReader, "beam_pol");
   TTreeReaderValue<double> target_pol(dataReader, "target_pol");
   TTreeReaderValue<double> phi(dataReader, "phi");
-  TTreeReaderValue<double> currentVariable(dataReader, propertyNames[currentFits].c_str());
+  // TTreeReaderValue<double> currentVariable(dataReader, propertyNames[currentFits].c_str());
+  TTreeReaderValue<int> currentVariable(dataReader, propertyNames[currentFits].c_str());
   // Counter to limit the number of processed entries
   while (dataReader.Next()) {
